refactor: split option utilities by responsibility
